08/20/24 - Changed Marginal Tax Rate to be Effective Tax Rate
The Marginal Tax Rate input was changed to be an Effective Tax Rate input. This is because the term "effective tax rate" better describes the input.

All variable names, comments, and text referring to the Marginal Tax Rate input were changed to refer to an Effective Tax Rate input. This was done for both the Web and Excel versions.

The default value for effective tax rate is now set to 19. The default is set both in index.html and stock_calc.js. I eyeballed several effective tax rates and thought 19 was a better default value.

This is a summary of changes I made:

1. In index.html, stock_calc.js, and web_style.css "id_input__marginal_tax_rate" is now referred to as "id_input__effective_tax_rate".

2. In stock_calc.js, "marg_tax_rate" is now called "eff_tax_rate".

3. All comments in stock_calc.js referring to a "marginal" tax rate now refer to an "effective" tax rate.

4. The default values for the effective tax rate (formerly marginal tax rate) in stock_calc.js and index.html is now 19 instead of 15. I removed text from index.html describing the tax rate as an average corporate rate. I don't have support for this being an average.

5. In the Excel version, the text "Marginal Tax Rate" was replaced by "Effective Tax Rate". The default value is now 19 instead of 15.