08/28/24 - Added Annual Augmented Payout Ratio
I added a field to the calculator control panel called "Annual Augmented Payout Ratio". This was added to both versions: Web and Excel.

This allows firms to be better valued. Many firms return much of the equity to shareholders through buybacks or dividends. I did not account for these returns of capital. Instead, all capital was previously assumed to remain in the business and accumulate in equity. This caused residual income values for each year (and predicted stock price) to be lower.

If net income is positive, a portion of net income is spent on buybacks or dividends each year (augmented payout ratio). Next year's book value only increases by the amount of net income that is leftover after dividends or buybacks.

If net income is negative, buybacks and dividends are assumed to not occur (since the firm is unprofitable). The entire loss is subtracted from book value to get next year's book value.