09/27/24 - Changed Values of Class and ID Attributes
This change affects index.html, stock_calc.js, and web_style.css. Double underscores were removed from the values of class attributes (prefixed with "cl_") and id attributes (prefixed with "id_").

For example, "id_table__calculator_table" was changed to "id_table_calculator_table".

Additionally, the word "style" was removed from the value of class attributes. It didn't convey any information. For example, the value "cl_style__button_black" was changed to "cl_button_black".

The first word ("cl" or "id") refers to the attribute the value is associated with. The second word refers to which HTML element the value is associated with. Anything after the second word is naming information. In value "cl_button_black", the value is associated with a class attribute ("cl"), used for a button, and the word "black" is the name.